Create ResourceAllocation package and finish FactoryGraphEvaluationService

+++ src/main/java/org/chainoptim/features/factory/service/FactoryInventoryService.java
@@ -27,5 +27,5 @@ public interface FactoryInventoryService {
     FactoryInventoryItem updateFactoryInventoryItem(UpdateFactoryInventoryItemDTO itemDTO);
 
     // Delete
-    void deleteFactoryInventoryItem(Long itemId);
+    void deleteFactoryInventoryItem(Integer itemId);
 }
